+++ src/main/java/org/codeforamerica/shiba/output/caf/TotalIncomeCalculator.java
@@ -4,14 +4,14 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class TotalIncomeCalculator {
-    public Double calculate(TotalIncome totalIncome) {
+    public int calculate(TotalIncome totalIncome) {
         if (totalIncome.getJobIncomeInformationList().isEmpty()) {
             return totalIncome.getLast30DaysIncome();
         } else {
             return totalIncome.getJobIncomeInformationList().stream().reduce(
-                    0.0,
+                    0,
                     (total, jobIncomeInfo) -> total + jobIncomeInfo.grossMonthlyIncome(),
-                    Double::sum
+                    Integer::sum
             );
         }
     }
